Refactor (split converter into "to" and "from" converters)
